+++ src/net.cpp
@@ -11,6 +11,7 @@
 #include <iostream>
 #include <fstream>
 #include "net.h"
+#include "firewall.cpp"
 #include <string>
 #include "addrman.h"
 #include "chainparams.h"
@@ -110,592 +111,6 @@ static list<CNode*> vNodesDisconnected;
 static CNodeSignals g_signals;
 CNodeSignals& GetNodeSignals() { return g_signals; }
 // **************************************
-
-// $$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$
-// [Bitcoin Firewall 1.2.2
-//  Aug 22, 2017 - Biznatch Enterprises & Bata Development
-// http://bata.io & https://github.com/BiznatchEnterprises/BitcoinFirewall
-
-string ModuleName = "[Bitcoin Firewall 1.2.2]";
-//boost::filesystem::path pathFirewallLog = GetDataDir()  / "firewall.log";
-
-// * FireWall Controls *
-bool ENABLE_FIREWALL = true;
-bool LIVE_DEBUG_OUTPUT = false;
-bool DETECT_INVALID_WALLET = true;
-bool BLACKLIST_INVALID_WALLET = true;
-bool BAN_INVALID_WALLET = true;
-bool DETECT_BANDWIDTH_ABUSE =  true;
-bool BLACKLIST_BANDWIDTH_ABUSE = true;
-bool BAN_BANDWIDTH_ABUSE = true;
-bool FALSE_POSITIVE_PROTECTION =  true;
-bool FIREWALL_CLEAR_BANS = false;
-
-// * Global Firewall Variables *
-int CurrentAverageHeight = 0;
-int CurrentAverageHeight_Min = 0;
-int CurrentAverageHeight_Max = 0;
-double CurrentAverageTraffic = 0;
-double CurrentAverageTraffic_Min = 0;
-double CurrentAverageTraffic_Max = 0;
-int CurrentAverageSend = 0;
-int CurrentAverageRecv = 0;
-int ALL_CHECK_TIMER = GetTime();
-int ALL_CHECK_MAX = 3;  // minutes interval
-int MINIMUM_PROTOCOL = 80007;
-// Not Used: int CurrentAverageHeight_Max = 0;
-// * BlackList node/peers Array
-string BLACKLIST[256];
-bool BLACKLIST_ATTACK = false;  // global temp variables
-bool BAN_ATTACK = false; //global temp variables
-int BlackListCounter = 0;
-string IgnoreSeedNode = "68.71.58.226:5784";  // WHITELIST (ignore)
-// * Attack Detection Settings *
-int AverageTolerance = 2;    // Reduce for minimal fluctuation 2 Blocks tolerance
-int AverageRange = 500;   // + or - Starting Height Range
-/// Bandwidth monitoring ranges
-double TrafficTolerance = 0.0001; // Reduce for minimal fluctuation
-double TrafficZone = 4; // + or - Traffic Range
-
-
-bool AddToBlackList(CNode *pnode)
-{
-// [AddTo Blacklist]
-
-        // Restart Blacklist count
-        if (BlackListCounter >  255)
-        {
-            BlackListCounter = 0;
-        }
-
-        BLACKLIST[BlackListCounter] = pnode->addrName;
-
-        // increase Blacklist count
-        BlackListCounter = BlackListCounter + 1;
-        // Add node IP to blacklist
-        //BlackList[BlackListCounter] = pnode->addrName;
-        pnode->nBlacklisted = pnode->nBlacklisted + 1;
-        // Append Blacklist to debug.log
-        LogPrintStr(ModuleName + " - Blacklisted: " + pnode->addrName.c_str() + "\n");
-return true;
-
-}
-
-
-bool ForceDisconnectNode(CNode *pnode, string FromFunction)
-{
-
-    // [Force Disconnection of node/peer]
-    //
-    //      Hard-disconnection function (Panic)
-
-
-TRY_LOCK(pnode->cs_vSend, lockSend);
-if (lockSend){
-    // release outbound grant (if any)
-    pnode->CloseSocketDisconnect();
-    LogPrintStr(ModuleName + " - (" + FromFunction + ") Panic Disconnect: " + pnode->addrName.c_str() + "\n");
-    return true;  
-}
-else
-{
-    pnode->vSendMsg.end();
-}
-
-return false;
-}
-
-
-bool CheckBlackList(CNode *pnode)
-{
-
-    if (pnode->nBlacklisted > 0)
-    {
-        // Banned IP FOUND!
-        return true;
-    }
-
-    for (int i = 0; i < BlackListCounter - 1; i++)
-    {  
-        if (pnode->addrName == BLACKLIST[i])
-        {   
-            // Banned IP FOUND!
-            return true;
-        }
-
-    }
-
-    // Banned IP not found
-    return false;
-}
-
-
-bool CheckAttack(CNode *pnode)
-{
-// [Intelligent Attack Protection]
-
-    bool DETECTED = false;
-    int nTimeConnected = GetTime() - pnode->nTimeConnected;
-    string AttackType = "";
-
-
-    // ---Filter 1 -------------
-    if (DETECT_BANDWIDTH_ABUSE == true)
-    {
-    // * Attack detection #2 & 3
-        // Calculate the ratio between Recieved bytes and Sent Bytes
-        // Detect a valid syncronizaion vs. a flood attack
-        
-        if (nTimeConnected > 10)
-        {
-            // * Attack detection #2
-            // Node is further ahead on the chain than average minimum
-            if (pnode->nStartingHeight > CurrentAverageHeight_Min)
-            {
-                if (pnode->nTrafficAverage < CurrentAverageTraffic_Min)
-                {
-                        // too low bandiwidth ratio limits
-                        DETECTED = true;
-                        AttackType = "2-LowBW-HighHeight";
-                }
-
-                if (pnode->nTrafficAverage > CurrentAverageTraffic_Max)
-                {
-                        // too high bandiwidth ratio limits
-                        DETECTED = true;
-                        AttackType = "2-HighBW-HighHeight";
-                }
-            }
-
-            // * Attack detection #3
-            // Node is behind on the chain than average minimum
-            if (pnode->nStartingHeight < CurrentAverageHeight_Min)
-            {  
-                if (pnode->nTrafficAverage < CurrentAverageTraffic_Min)
-                {
-                        // too low bandiwidth ratio limits
-                        DETECTED = true;
-                        AttackType = "3-LowBW-LowHeight";
-                }
-
-                if (pnode->nTrafficAverage > CurrentAverageTraffic_Max)
-                {
-
-                    // too high bandiwidth ratio limits
-                    DETECTED = true;
-                    AttackType = "3-HighBW-LowHeight";
-                }
-
-            }
-
-        }
-
-        if (DETECTED == true)
-        {
-            if (BLACKLIST_BANDWIDTH_ABUSE == true)
-            {
-                BLACKLIST_ATTACK = true;
-            }
-
-            if (BAN_BANDWIDTH_ABUSE == true)
-            {
-                BAN_ATTACK = true;
-            }
-
-        }
-    }
-    // ----------------
-
-    if (FALSE_POSITIVE_PROTECTION == true)
-    {
-        // ----------------
-        // FALSE POSITIVE PROTECTION
-        if (DETECTED == true)
-        {
-
-            if (AttackType == "2-LowBW-HighHeight")
-            {
-                AttackType = "";
-                DETECTED = false;
-            }   
-
-            if (AttackType == "2-HighBW-HighHeight")
-            {
-                // Node/peer is in wallet sync (catching up to full blockheight)
-                AttackType = "";
-                DETECTED = false;
-            }
-
-            if (AttackType == "3-LowBW-LowHeight")
-            {
-                AttackType = "";
-                DETECTED = false;
-            }   
-
-           if (AttackType == "3-HighBW-LowHeight")
-            {
-                double tnTraffic = pnode->nSendBytes / pnode->nRecvBytes;
-                if (pnode->nTrafficAverage < CurrentAverageTraffic_Max)
-                {
-                    if (tnTraffic > 17.2 || tnTraffic < 17.1)
-                    {
-                        // wallet full sync
-                        AttackType = "";
-                        DETECTED = false;
-                    }
-                }
-
-                if (pnode->nSendBytes < pnode->nRecvBytes)
-                {
-                    // wallet full sync
-                    AttackType = "";
-                    DETECTED = false;
-                }
-            }   
-        }
-    }
-    // ----------------
-
-    // ---Filter 2-------------
-    if (DETECT_INVALID_WALLET == true)
-    {
-    // * Attack detection #1
-        // (Start Height = -1, over 30 seconds connection length)
-        // Check for more than 600 seconds connection length
-        if (nTimeConnected > 90)
-        {
-            // Check for -1 blockheight
-            if (pnode->nStartingHeight == -1)
-            {
-                // Trigger Blacklisting
-                DETECTED = true;
-                AttackType = "1-StartHeight-Invalid";
-            }
-        }
-
-        // Check for -1 blockheight
-        if (nTimeConnected > 90)
-        {
-            // Check for -1 blockheight
-            if (pnode->nStartingHeight < 0)
-            {
-                // Trigger Blacklisting
-                DETECTED = true;
-                AttackType = "1-StartHeight-Invalid";
-            }
-        }
-        
-        // (Protocol: 0
-        // Check for more than 30 seconds connection length
-        if (nTimeConnected > 90)
-        {
-            // Check for 0 protocol
-            if (pnode->nRecvVersion == 0)
-            {
-                // Trigger Blacklisting
-                DETECTED = true;
-                AttackType = "1-Protocol-Invalid";
-
-            }
-        }
-
-        // (Protocol: 0
-        // Check for more than 30 seconds connection length
-        if (nTimeConnected > 90)
-        {
-            // Check for 
-            if (pnode->nRecvVersion < 1)
-            {
-                // Trigger Blacklisting
-                DETECTED = true;
-                AttackType = "1-Protocol-Invalid";
-
-            }
-        }
-
-        if (DETECTED == false)
-        {
-            int GlobalTimeOn = GetTime() - ALL_CHECK_TIMER;
-            if (GlobalTimeOn > ALL_CHECK_MAX * 60)
-            {
-                ALL_CHECK_TIMER = GetTime();
-
-                if (LIVE_DEBUG_OUTPUT == true)
-                {
-                    cout<<ModuleName<<" - Random Check"<<endl;
-                }
-
-                BOOST_FOREACH(CNode* pnode, vNodes)
-                {
-                    int tnTimeConnected = GetTime() - pnode->nTimeConnected;
-                    if (tnTimeConnected > 90)
-                    {
-                        DETECTED = false;
-                        AttackType = "";
-
-                        if (pnode->nStartingHeight < 0)
-                        {
-                            DETECTED = true;
-                            AttackType = "1-Invalid-Height";
-
-                        }
-
-                        if (pnode->nRecvVersion < MINIMUM_PROTOCOL)
-                        {
-                            DETECTED = true;
-                            AttackType = "1-Obsolete-Version";
-                        }
-
-                        if (pnode->nStartingHeight < CurrentAverageHeight_Min)
-                        {
-                            if (pnode->nTrafficAverage > CurrentAverageTraffic_Max)
-                            {
-                                double tnTraffic = pnode->nSendBytes / pnode->nRecvBytes;
-                                if (tnTraffic > 17.1)
-                                {
-                                    if (tnTraffic < 17.12)
-                                    {
-                                    // Double Spend
-                                    AttackType = "1-Double-Spend";
-                                    DETECTED = true;
-                                    }
-                                }
-
-                            }
-
-                            if (tnTimeConnected > 200)
-                            {
-                                if (tnTimeConnected < 600)
-                                {
-
-                                    if (pnode->nSendBytes == pnode->nRecvBytes)
-                                    {
-                                        // Double Spend
-                                        AttackType = "1-Double-Spend";
-                                        DETECTED = true;  
-                                    }
-
-                                }
-                            }
-
-
-                            if (tnTimeConnected < 800)
-                            {
-                                if (pnode->nSendBytes > 40000)
-                                {   
-                                    if (pnode->nRecvBytes > 80000)
-                                    {
-                                        // Double Spend
-                                        AttackType = "1-Double-Spend";
-                                        DETECTED = true;  
-                                    }
-                                }
-                            }
-                        }
-
-                        if (LIVE_DEBUG_OUTPUT == true)
-                        {
-                            cout<<ModuleName<<" [IP: "<<pnode->addrName.c_str()<<"] [Protocol: "<<pnode->nRecvVersion<<"] [Send: "<<pnode->nSendBytes<<"] [Recv: "<<pnode->nRecvBytes<<"] Detected: "<<DETECTED<<"]"<<endl;
-                        }
-
-                        if (DETECTED == true)
-                        {
-
-                            std::string NodeTrafficRatioStr = boost::lexical_cast<std::string>(pnode->nTrafficRatio);
-                            std::string NodeTrafficAverageStr = boost::lexical_cast<std::string>(pnode->nTrafficAverage);
-                            std::string CurrentAverageTrafficStr = boost::lexical_cast<std::string>(CurrentAverageTraffic);
-                            std::string SendBytesStr = boost::lexical_cast<std::string>(pnode->nSendBytes);
-                            std::string RecvBytesStr = boost::lexical_cast<std::string>(pnode->nRecvBytes);
-
-                            LogPrintStr(ModuleName + " - [Attack Type: " +  AttackType + "] [Detected from: " + pnode->addrName.c_str() + "] [Node Traffic: " + NodeTrafficRatioStr +  "] [Node Traffic Avrg: " + NodeTrafficAverageStr + "] [Traffic Avrg: " + CurrentAverageTrafficStr + "] [Sent Bytes: " + SendBytesStr + "] [Recv Bytes: " + RecvBytesStr + "]\n");
-
-                            if (BLACKLIST_INVALID_WALLET == true)
-                            {   
-                            AddToBlackList(pnode);
-                            }
-                            ForceDisconnectNode(pnode, "Random");
-
-                        }
-               
-                    }
-
-                }
-
-            }
-
-        }
-
-
-        if (DETECTED == true)
-        {
-            if (BLACKLIST_INVALID_WALLET == true)
-            {
-                BLACKLIST_ATTACK = true;
-            }
-
-            if (BAN_INVALID_WALLET == true)
-            {
-                BAN_ATTACK = true;
-            }
-
-        }
-
-    }
-
-
-    // ----------------
-    // ATTACK DETECTED (TRIGGER)!
-    if (DETECTED == true)
-    {
-
-        std::string NodeTrafficRatioStr = boost::lexical_cast<std::string>(pnode->nTrafficRatio);
-        std::string NodeTrafficAverageStr = boost::lexical_cast<std::string>(pnode->nTrafficAverage);
-        std::string CurrentAverageTrafficStr = boost::lexical_cast<std::string>(CurrentAverageTraffic);
-        std::string SendBytesStr = boost::lexical_cast<std::string>(pnode->nSendBytes);
-        std::string RecvBytesStr = boost::lexical_cast<std::string>(pnode->nRecvBytes);
-
-        LogPrintStr(ModuleName + " - [Attack Type: " +  AttackType + "] [Detected from: " + pnode->addrName.c_str() + "] [Node Traffic: " + NodeTrafficRatioStr +  "] [Node Traffic Avrg: " + NodeTrafficAverageStr + "] [Traffic Avrg: " + CurrentAverageTrafficStr + "] [Sent Bytes: " + SendBytesStr + "] [Recv Bytes: " + RecvBytesStr + "]\n");
-
-        // Blacklist IP on Attack detection
-        // * add node/peer IP to blacklist
-        if (BLACKLIST_ATTACK == true)
-        {
-            AddToBlackList(pnode);
-        }
-
-        // Peer/Node Ban if required
-        if (BAN_ATTACK == true)
-        {
-            pnode->Ban(pnode->addr);
-            LogPrintStr(ModuleName + " - Banned: " + pnode->addrName.c_str() + "\n");
-        }
-
-// ATTACK DETECTED!
-return true;
-    }
-    else
-    {
-//NO ATTACK DETECTED...
-return false;
-    }
-    // ----------------
-}
-
-
-void Examination(CNode *pnode, string FromFunction)
-{
-// Calculate new Height Average from all peers connected
-
-    bool UpdateNodeStats = false;
-
-    if (CheckAttack(pnode) == true)
-    { 
-        // Peer/Node Panic Disconnect
-        ForceDisconnectNode(pnode, FromFunction);
-    }
-
-    // ** Update current average if increased ****
-    if (pnode->nStartingHeight > CurrentAverageHeight) 
-    {
-        CurrentAverageHeight = CurrentAverageHeight + pnode->nStartingHeight; 
-        CurrentAverageHeight = CurrentAverageHeight / 2;
-        CurrentAverageHeight = CurrentAverageHeight - AverageTolerance;      // reduce with tolerance
-        CurrentAverageHeight_Min = CurrentAverageHeight - AverageRange;
-        CurrentAverageHeight_Max = CurrentAverageHeight + AverageRange;
-    }
-
-    if (pnode->nRecvBytes > 0)
-    {
-        pnode->nTrafficRatio = pnode->nSendBytes / (double)pnode->nRecvBytes;
-
-        if (pnode->nTrafficTimestamp == 0)
-        {
-            UpdateNodeStats = true;
-        }
-
-        if (GetTime() - pnode->nTrafficTimestamp > 20){
-            UpdateNodeStats = true;
-        }
-
-            pnode->nTrafficAverage = pnode->nTrafficAverage + (double)pnode->nTrafficRatio / 2;
-            pnode->nTrafficTimestamp = GetTime();
-
-        if (UpdateNodeStats == true)
-        {   
-            CurrentAverageTraffic = CurrentAverageTraffic + (double)pnode->nTrafficAverage;
-            CurrentAverageTraffic = CurrentAverageTraffic / (double)2;
-            CurrentAverageTraffic = CurrentAverageTraffic - (double)TrafficTolerance;      // reduce with tolerance
-            CurrentAverageTraffic_Min = CurrentAverageTraffic - (double)TrafficZone;
-            CurrentAverageTraffic_Max = CurrentAverageTraffic + (double)TrafficZone;    
-            CurrentAverageSend = CurrentAverageSend + pnode->nSendBytes / vNodes.size();
-            CurrentAverageRecv = CurrentAverageRecv + pnode->nRecvBytes / vNodes.size();
-            
-
-            //std::ofstream fout(pathFirewallLog);
-
-            if (LIVE_DEBUG_OUTPUT == true){
-            cout<<ModuleName<<" [BlackListed: "<< BlackListCounter<<"]"<<endl;
-            cout<<"[Traffic: "<<CurrentAverageTraffic<<"] [Traffic Min: "<<CurrentAverageTraffic_Min<<"] [Traffic Max: "<<CurrentAverageTraffic_Max<<"]"<<" [Height: "<<CurrentAverageHeight<<"] [Height Min: "<<CurrentAverageHeight_Min<<"] [Height Max: "<<CurrentAverageHeight_Max<<"] [Send Avrg: "<<CurrentAverageSend<<"] [Rec Avrg: "<<CurrentAverageRecv<<"]"<<endl;
-            cout<<"[Check Node IP: "<<pnode->addrName.c_str()<<"] [Traffic: "<<pnode->nTrafficRatio<<"] [Traffic Average: "<<pnode->nTrafficAverage<<"] [Starting Height: "<<pnode->nStartingHeight<<"] [Node Sent: "<<pnode->nSendBytes<<"] [Node Recv: "<<pnode->nRecvBytes<<"]"<<endl;
-            }
-
-        }
-
-    }
-
-
-
-}
-
-
-bool FireWall(CNode *pnode, string FromFunction)
-{
-
-    if (ENABLE_FIREWALL == false)
-    {
-        return false;
-    }
-
-    // Check for Static Whitelisted Seed Node
-    if (pnode->addrName == IgnoreSeedNode)
-    {
-        return false;
-    }
-
-    // Check for Node Whitelisted status
-    if (pnode->fWhitelisted == true)
-    {
-        return false;
-    }
-
-    if (FIREWALL_CLEAR_BANS == true)
-    {
-        pnode->ClearBanned();
-    }
-
-    // Check for 0 peer count (auto-unban)
-    if (vNodes.size() == 0){
-        pnode->ClearBanned();
-    }
-
-
-    if (CheckBlackList(pnode) == true)
-    {
-        FromFunction = "Blacklisted";
-
-// Peer/Node Panic Disconnect
-ForceDisconnectNode(pnode, FromFunction);
-return true;
-    }
-
-    // Perform a Node consensus examination
-Examination(pnode, FromFunction);
-
-// Peer/Node Safe    
-return false;
-
-}
-// $$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$
 
 void AddOneShot(string strDest)
 {
@@ -2768,7 +2183,6 @@ CNode::CNode(SOCKET hSocketIn, CAddress addrIn, std::string addrNameIn, bool fIn
     nVersion = 0;
     strSubVer = "";
     fWhitelisted = false;
-    nBlacklisted = 0;
     fOneShot = false;
     fClient = false; // set by version message
     fInbound = fInboundIn;
@@ -2780,6 +2194,7 @@ CNode::CNode(SOCKET hSocketIn, CAddress addrIn, std::string addrNameIn, bool fIn
     nSendOffset = 0;
     hashContinue = 0;
     nStartingHeight = -1;
+    nSyncHeight = 0;
     fGetAddr = false;
     fRelayTxes = false;
     setInventoryKnown.max_size(SendBufferSize() / 1000);
