+++ src/net.h
@@ -239,7 +239,8 @@ public:
     CCriticalSection cs_vRecvMsg;
     uint64_t nRecvBytes;
     int nRecvVersion;
-
+    int nSyncHeight;
+    int nSyncHeightOld;
     int64_t nLastSend;
     int64_t nLastRecv;
     int64_t nTimeConnected;
@@ -256,7 +257,6 @@ public:
     // the network or wire types and the cleaned string used when displayed or logged.
     std::string strSubVer, cleanSubVer;
     bool fWhitelisted; // This peer can bypass DoS banning & firewall.
-    int nBlacklisted; // This peer can't bypass firewall
     bool fOneShot;
     bool fClient;
     bool fInbound;
